+++ msmbuilder/tests/workflows/test_1.sh
@@ -1,8 +1,7 @@
-msmb AlanineDipeptide
-msmb AtomIndices --out atom_indices.txt -p ~/msmbuilder_data/alanine_dipeptide/ala2.pdb \
-     -d --heavy
-msmb AtomPairsFeaturizer --transformed atom_pairs --trjs '~/msmbuilder_data/alanine_dipeptide/*.dcd' \
-    --pair_indices atom_indices.txt --top ~/msmbuilder_data/alanine_dipeptide/ala2.pdb --out atom_pairs.pkl
+msmb AlanineDipeptide --data_home ./
+msmb AtomIndices --out atom_indices.txt -p ./alanine_dipeptide/ala2.pdb -d --heavy
+msmb AtomPairsFeaturizer --transformed atom_pairs --trjs './alanine_dipeptide/*.dcd' \
+    --pair_indices atom_indices.txt --top ./alanine_dipeptide/ala2.pdb --out atom_pairs.pkl
 msmb tICA -i atom_pairs/ -t atom_pairs_tica.h5  --n_components 4 \
     --shrinkage 0 \
     --kinetic_mapping \
